# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bodypower Gym API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  /** @nullable */
  membershipStatus?: string | null;
  /** @nullable */
  membershipExpiry?: string | null;
  language?: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  name?: string;
  phone?: string;
  language?: string;
}

export interface PasswordChangeInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface MembershipPlan {
  id: string;
  name: string;
  price: number;
  duration: string;
  features: string[];
  popular?: boolean;
}

export type BookingInputMembershipType = typeof BookingInputMembershipType[keyof typeof BookingInputMembershipType];


export const BookingInputMembershipType = {
  demo: 'demo',
  silver: 'silver',
  premium: 'premium',
} as const;

export interface BookingInput {
  fullName: string;
  phone: string;
  /**
     * @minimum 18
     * @maximum 60
     */
  age: number;
  email: string;
  membershipType: BookingInputMembershipType;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface Payment {
  id: number;
  bookingId: number;
  amount: number;
  method: string;
  status: PaymentStatus;
  /** @nullable */
  transactionId?: string | null;
  createdAt: string;
}

export interface Booking {
  id: number;
  userId: number;
  fullName: string;
  phone: string;
  age: number;
  email: string;
  membershipType: string;
  status: BookingStatus;
  /** @nullable */
  confirmationCode?: string | null;
  createdAt: string;
  payment?: Payment;
}

export type PaymentInputMethod = typeof PaymentInputMethod[keyof typeof PaymentInputMethod];


export const PaymentInputMethod = {
  upi: 'upi',
  credit_card: 'credit_card',
  debit_card: 'debit_card',
  net_banking: 'net_banking',
} as const;

export interface PaymentInput {
  bookingId: number;
  amount: number;
  method: PaymentInputMethod;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  cardNumber?: string | null;
  /** @nullable */
  cardHolder?: string | null;
  /** @nullable */
  cardExpiry?: string | null;
  /** @nullable */
  cardCvv?: string | null;
  /** @nullable */
  bankName?: string | null;
}

export interface DashboardStats {
  totalBookings: number;
  activeMembers: number;
  totalSpent: number;
  upcomingBookings: Booking[];
  recentPayments?: Payment[];
}

export interface BookingTypeStat {
  type: string;
  count: number;
}

export interface MonthRevenue {
  month: string;
  revenue: number;
}

export interface PaymentMethodStat {
  method: string;
  count: number;
  total: number;
}

export interface AdminStats {
  totalUsers: number;
  totalBookings: number;
  totalRevenue: number;
  activeMemberships: number;
  bookingsByType?: BookingTypeStat[];
  revenueByMonth?: MonthRevenue[];
  paymentsByMethod?: PaymentMethodStat[];
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  membershipStatus?: string | null;
  totalBookings?: number;
  totalSpent?: number;
  createdAt: string;
}