# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bodypower Gym API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { Payment } from './payment';

export interface Booking {
  id: number;
  userId: number;
  fullName: string;
  phone: string;
  age: number;
  email: string;
  membershipType: string;
  status: BookingStatus;
  /** @nullable */
  confirmationCode?: string | null;
  createdAt: string;
  payment?: Payment;
}