# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bodypower Gym API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputMembershipType } from './bookingInputMembershipType';

export interface BookingInput {
  fullName: string;
  phone: string;
  /**
     * @minimum 18
     * @maximum 60
     */
  age: number;
  email: string;
  membershipType: BookingInputMembershipType;
}