# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bodypower Gym API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipPlan {
  id: string;
  name: string;
  price: number;
  duration: string;
  features: string[];
  popular?: boolean;
}